in-progress - working without tokens

+++ app/controllers/registrations_controller.rb
@@ -15,10 +15,34 @@ class RegistrationsController < ApplicationController
     season_id = params[:season]
     season = season_id.to_i.is_a?(Numeric) ? (@club.seasons.accepting_registrations_now.where(:id => season_id.to_i).first.nil? ? nil : Season.find_by_id(@club.seasons.accepting_registrations_now.where(:id => season_id.to_i).first.id)) : nil
     if season.present?
-      @registration = Registration.new(:club => @club, :season => season, :player_attributes => {:birthdate => Date.civil(Date.today.years_ago(5).year, 1, 1), :person_attributes => @player_person_defaults})
-      RegistrationQuestionResponse.populate_responses_for_registration(@registration, true)
-      @registration.registrations_people.build(:person => Person.new(@person_defaults), :person_role => PersonRole.find_by_role_abbreviation('PG'), :primary => true)
-      @registration.registrations_people.build(:person => Person.new(@person_defaults), :person_role => PersonRole.find_by_role_abbreviation('PG'), :primary => false)
+      @registration = nil
+      player_extid = params[:player]
+      # is there a player param, and does it return a valid player for this club?
+      if player_extid.present? && Player.find_by_extid(player_extid).present?
+        player = Player.find_by_extid(player_extid)
+    
+        @registration = Registration.new(:club => @club, :season => season, :player => player)
+          RegistrationQuestionResponse.populate_responses_for_registration(@registration, true)
+
+        prev_pg = player.registrations.last.registrations_people.parent_guardians
+        
+        prev_pg.each do |rp|
+          @registration.registrations_people.build(:person => rp.person, :person_role => rp.person_role, :primary => rp.primary)
+        end
+        
+        if prev_pg.count == 0
+          @registration.registrations_people.build(:person => Person.new(@person_defaults), :person_role => PersonRole.find_by_role_abbreviation('PG'), :primary => true)
+        end
+        if prev_pg.count < 2
+          @registration.registrations_people.build(:person => Person.new(@person_defaults), :person_role => PersonRole.find_by_role_abbreviation('PG'), :primary => false)                    
+        end
+        
+      else
+        @registration = Registration.new(:club => @club, :season => season, :player_attributes => {:birthdate => Date.civil(Date.today.years_ago(5).year, 1, 1), :person_attributes => @player_person_defaults})
+        RegistrationQuestionResponse.populate_responses_for_registration(@registration, true)
+        @registration.registrations_people.build(:person => Person.new(@person_defaults), :person_role => PersonRole.find_by_role_abbreviation('PG'), :primary => true)
+        @registration.registrations_people.build(:person => Person.new(@person_defaults), :person_role => PersonRole.find_by_role_abbreviation('PG'), :primary => false)
+      end
     else
       redirect_to club_url(@club.subdomain)
     end
@@ -26,13 +50,46 @@ class RegistrationsController < ApplicationController
 
   def create    
     reg_params = params[:registration]
-    
+
     # clean out parent/guardian 2 if they are empty
-    if reg_params[:registrations_people_attributes]["1"][:first_name].blank? && reg_params[:registrations_people_attributes]["1"][:last_name].blank?
+    if reg_params[:registrations_people_attributes]["1"].present? && reg_params[:registrations_people_attributes]["1"][:person_attributes][:first_name].blank? && reg_params[:registrations_people_attributes]["1"][:person_attributes][:last_name].blank?
       reg_params[:registrations_people_attributes].delete("1")
     end
+        
+    if reg_params[:player_attributes][:id].blank?
+      @registration = @club.registrations.new(reg_params)
+    else
+      # returning player reg, so we need to manually create the
+      #   registrations_person records because of weak-ass rails
+      #   nested attribute handling. Bah humbug!
+      primary_guardian_person_id = reg_params[:registrations_people_attributes]["0"][:person_attributes][:id]
+      secondary_guardian_person_id = reg_params[:registrations_people_attributes]["1"][:person_attributes][:id] if reg_params[:registrations_people_attributes]["1"].present?
       
-    @registration = @club.registrations.new(reg_params)
+      registrations_people = []
+      primary_guardina = nil
+      secondary_guardian = nil
+      if primary_guardian_person_id.present?
+        primary_guardian = RegistrationsPerson.new(:person_id => primary_guardian_person_id, :person_role => PersonRole.find_by_role_abbreviation('PG'), :primary => true)
+        primary_guardian.attributes = reg_params[:registrations_people_attributes]["0"]
+      else
+        primary_guardian = RegistrationsPerson.new(reg_params[:registrations_people_attributes]["0"])
+      end
+      registrations_people << primary_guardian
+      if reg_params[:registrations_people_attributes]["1"].present?
+        if secondary_guardian_person_id.present?
+          secondary_guardian = RegistrationsPerson.new(:person_id => secondary_guardian_person_id, :person_role => PersonRole.find_by_role_abbreviation('PG'), :primary => false)
+          secondary_guardian.attributes = reg_params[:registrations_people_attributes]["1"]
+        else
+          secondary_guardian = RegistrationsPerson.new(reg_params[:registrations_people_attributes]["1"])
+        end        
+        registrations_people << secondary_guardian
+      end
+      reg_params.delete(:registrations_people_attributes)
+            
+      @registration = @club.registrations.new(:player_id => reg_params[:player_attributes][:id].to_i, :registrations_people => registrations_people)      
+      @registration.attributes = reg_params
+    end
+    
     season = @club.seasons.accepting_registrations_now.where(:id => @registration.season_id).first
     unless season.present?
       redirect_to club_url(@club.subdomain)
@@ -58,6 +115,8 @@ class RegistrationsController < ApplicationController
           flash.now[:error] = @message
           render :action => "new"      
         else
+          # don't want someone to try and hack this registration into a different club, now do we!
+          @registration.update_attribute(:club_id, @club.id)
           @pp = PaymentPackage.for_season_and_division(@registration.season, @registration.division)
           RegistrationQuestionResponse.populate_responses_for_registration(@registration, false)
           RegistrationQuestionResponse.create_default_responses_for_protected_questions(@registration)
@@ -99,11 +158,15 @@ class RegistrationsController < ApplicationController
     reg = @club.registrations.find(reg_id)
 
     reg.registrations_people.each do |p|
-      p.person.destroy
+      unless p.person.registrations_people.count > 1
+        p.person.destroy
+      end
       p.destroy
     end
-    reg.player.person.destroy
-    reg.player.destroy
+    unless reg.player.registrations.count > 1
+      reg.player.person.destroy
+      reg.player.destroy
+    end
     reg.destroy
     
     redirect_to regreport_url(@club.subdomain)
